working on update/delete

+++ src/pages/question-list/question-list.ts
@@ -1,20 +1,22 @@
-import { Component } from '@angular/core';
-import { Questionsform } from '../questionsform/questionsform'
+import { Component, Output, EventEmitter } from '@angular/core';
+// import { Questionsform } from '../questionsform/questionsform'
 import { Xapi } from '../../xmodule/providers/xapi';
 import * as xi from '../../xmodule/interfaces/xapi';
-import { QuestionForm } from '../../providers/share';
+import { PostQuery } from '../../providers/share';
 import { Dashboard } from '../dashboard/dashboard';
-import { NavController } from 'ionic-angular';
+import { NavController, AlertController } from 'ionic-angular';
 @Component( {
     templateUrl: `question-list.html`
 })
 export class QuestionList {
-    questionForm : QuestionForm = <QuestionForm> {};
+
+    @Output() edit = new EventEmitter<number>();
+
     posts: xi.PostQueryResponse;
-    constructor( private x: Xapi, private navCtrl: NavController ) {
+    constructor( private x: Xapi, private navCtrl: NavController, private alrtCtrl: AlertController) {
         console.log('Dashboard::onClickList()');
-        let args: xi.PostQuery = <xi.PostQuery> {};
-        args.category_name = 'question';
+        let args: PostQuery = <PostQuery> {};
+        args.category = 'question';
         this.x.get_posts( args, re => {
         console.log('callback: ', re);
         this.posts = re.data;
@@ -25,27 +27,42 @@ export class QuestionList {
     onClickBack(){
         console.log('clicked back');
         this.navCtrl.setRoot( Dashboard );
-        
+
     }
-    onClickUpdate(id){
-        this.navCtrl.setRoot( Questionsform,{
-            title:'Update',
-            id: id
-        } );
+    onClickUpdate(ID){
+        console.log('onClickEdit()', ID);
+        this.edit.emit( ID );
     }
 
     onClickDelete(ID){
         console.log(ID);
-        this.x.delete_post( ID,(res: xi.Response) => {
-            if(res.success){
-                delete this.posts[ res.data ];
-                this.x.alert('Delete','Question Deleted successfully')
-            }else{
-                this.x.error ( res.data.message );
+        let confirmDelete = this.alrtCtrl.create({
+          title: 'Confirmation',
+          subTitle:'Are you sure you want to delete this question',
+          buttons:[{
+            text:'Ok',
+            handler: ()=>{
+              console.log('deleteClicked OK');
+              this.x.delete_post( ID,(res: xi.Response) => {
+                if(res.success){
+                  delete this.posts[ res.data ];
+                  this.x.alert('Delete','Question Deleted successfully');
+                  this.navCtrl.push(this);
+                }else{
+                  this.x.error ( res.data.message );
+                }
+              },e=>{
+                this.x.error( e );
+              })
             }
-        },e=>{
-            this.x.error( e );
-        })
+          },{
+            text:'Cancel',
+            handler: ()=>{
+
+            }
+          }]
+        });
+      confirmDelete.present();
     }
 
 }
